?Renamed the cli and corresponding make target
