+++ tests/unit_test/refresh_data/page_views_and_downloads_monthly_cli_test.py
@@ -4,8 +4,8 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 
-from data_hub_metrics_api.refresh_data.page_views_monthly_cli import main
-import data_hub_metrics_api.refresh_data.page_views_monthly_cli as cli_module
+from data_hub_metrics_api.refresh_data.page_views_and_downloads_monthly_cli import main
+import data_hub_metrics_api.refresh_data.page_views_and_downloads_monthly_cli as cli_module
 
 
 @pytest.fixture(name="get_redis_client_mock", autouse=True)
